chore: rename file to comply standards
