+++ hrm-domain/hrm-core/config/ssmConfigurationCache.ts
