+++ hrm-domain/hrm-core/permissions/json-permissions.ts
@@ -14,7 +14,7 @@
  * along with this program.  If not, see https://www.gnu.org/licenses/.
  */
 
-import { getFromSSMCache } from '../config/ssm-cache';
+import { getFromSSMCache } from '../config/ssmConfigurationCache';
 import { rulesMap } from './rulesMap';
 import { Permissions } from './index';
 import { AccountSID } from '@tech-matters/types';
